Rename files and update imports
- Rename row_conditions.py -> legacy_row_conditions.py using git mv
- Rename conditions.py -> row_conditions.py using git mv
- Create new conditions.py as backwards compatibility re-export module
- Update all imports throughout codebase
