+++ great_expectations/expectations/core/expect_table_row_count_to_be_between.py
@@ -8,7 +8,6 @@ from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.suite_parameters import (
     SuiteParameterDict,  # noqa: TC001 # FIXME CoP
 )
-from great_expectations.expectations.conditions import RowConditionType  # noqa: TC001
 from great_expectations.expectations.expectation import (
     BatchExpectation,
     render_suite_parameter_string,
@@ -18,6 +17,7 @@ from great_expectations.expectations.model_field_descriptions import FAILURE_SEV
 from great_expectations.expectations.model_field_types import (
     ConditionParser,  # noqa: TC001 # FIXME CoP
 )
+from great_expectations.expectations.row_conditions import RowConditionType  # noqa: TC001
 from great_expectations.render import LegacyRendererType, RenderedStringTemplateContent
 from great_expectations.render.renderer.renderer import renderer
 from great_expectations.render.renderer_configuration import (
