+++ tests/integration/data_sources_and_expectations/test_expectation_conditions.py
@@ -9,8 +9,10 @@ from great_expectations.compatibility.postgresql import POSTGRESQL_TYPES
 from great_expectations.compatibility.snowflake import SNOWFLAKE_TYPES
 from great_expectations.compatibility.sqlalchemy import sqltypes
 from great_expectations.datasource.fluent.interfaces import Batch
-from great_expectations.expectations.conditions import PassThroughCondition
-from great_expectations.expectations.row_conditions import Column
+from great_expectations.expectations.row_conditions import (
+    Column,
+    PassThroughCondition,
+)
 from tests.integration.conftest import parameterize_batch_for_data_sources
 from tests.integration.test_utils.data_source_config import (
     BigQueryDatasourceTestConfig,
